requirements.txt
Package                      Version
---------------------------- ---------
absl-py                      1.4.0
astunparse                   1.6.3
cachetools                   5.3.0
certifi                      2022.12.7
charset-normalizer           3.0.1
cloudpickle                  2.2.1
contourpy                    1.0.7
cycler                       0.11.0
decorator                    5.1.1
dm-tree                      0.1.8
flatbuffers                  23.1.21
fonttools                    4.38.0
gast                         0.4.0
google-auth                  2.16.1
google-auth-oauthlib         0.4.6
google-pasta                 0.2.0
grpcio                       1.51.3
gym                          0.21.0
gym-notices                  0.0.8
h5py                         3.8.0
idna                         3.4
keras                        2.11.0
kiwisolver                   1.4.4
libclang                     15.0.6.1
Markdown                     3.4.1
MarkupSafe                   2.1.2
matplotlib                   3.7.0
numpy                        1.24.2
oauthlib                     3.2.2
opt-einsum                   3.3.0
packaging                    23.0
Pillow                       9.4.0
pip                          22.3.1
protobuf                     3.19.6
pyasn1                       0.4.8
pyasn1-modules               0.2.8
pyparsing                    3.0.9
python-dateutil              2.8.2
PyYAML                       6.0
requests                     2.28.2
requests-oauthlib            1.3.1
rsa                          4.9
scipy                        1.10.1
setuptools                   65.6.3
six                          1.16.0
tensorboard                  2.11.2
tensorboard-data-server      0.6.1
tensorboard-plugin-wit       1.8.1
tensorboardX                 2.6
tensorflow                   2.11.0
tensorflow-estimator         2.11.0
tensorflow-io-gcs-filesystem 0.30.0
tensorflow-probability       0.19.0
termcolor                    2.2.0
tqdm                         4.64.1
typing_extensions            4.5.0
urllib3                      1.26.14
Werkzeug                     2.2.3
wheel                        0.38.4
wrapt                        1.14.1
